requirements.txt
discord.py==2.3.2
google-generativeai==0.5.2
pillow==10.3.0
requests==2.32.2
python-dotenv==1.0.1
uvicorn==0.30.0
fastapi==0.110.1
beautifulsoup4==4.12.3
wikipedia==1.4.0
aiohttp==3.9.5
pydub==0.25.1